--- src/main/java/kuit/springbasic/controller/user/LogOutController.java
@@ -1,33 +1,0 @@
-package kuit.springbasic.controller.user;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
-import kuit.springbasic.db.UserRepository;
-import kuit.springbasic.domain.User;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Slf4j
-@Controller
-@RequiredArgsConstructor
-@RequestMapping("/user")
-// 세션 무효화만 진행
-public class LogOutController {
-
-    private final UserRepository userRepository;
-
-    @PostMapping("/logout")
-    public String logout(HttpServletRequest request) {
-        HttpSession session = request.getSession();
-        if(session != null) {
-            // 현재 세션의 모든 정보 삭제
-            session.invalidate();
-        }
-        log.info("logged out successfully");
-        return "redirect:/";
-    }
-}
